Renamed test to declare abstract nature
