+++ subprojects/platform-play/src/integTest/groovy/org/gradle/play/integtest/continuous/PlayReloadIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 the original author or authors.
+ * Copyright 2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,88 +16,142 @@
 
 package org.gradle.play.integtest.continuous
 
-import org.gradle.internal.filewatch.PendingChangesListener
-import org.gradle.internal.filewatch.PendingChangesManager
-import org.gradle.internal.filewatch.SingleFirePendingChangesListener
-import org.gradle.play.integtest.fixtures.AbstractMultiVersionPlayReloadIntegrationTest
-import org.gradle.play.integtest.fixtures.AdvancedRunningPlayApp
-import org.gradle.play.integtest.fixtures.PlayApp
-import org.gradle.play.integtest.fixtures.RunningPlayApp
-import org.gradle.play.integtest.fixtures.app.AdvancedPlayApp
-import org.gradle.test.fixtures.server.http.BlockingHttpServer
-import org.junit.Rule
+class PlayReloadIntegrationTest extends AbstractPlayReloadIntegrationTest {
 
-abstract class PlayReloadIntegrationTest extends AbstractMultiVersionPlayReloadIntegrationTest {
-    @Rule BlockingHttpServer server = new BlockingHttpServer()
+    def "should reload modified scala controller and routes"() {
+        when:
+        succeeds("runPlayBinary")
 
-    RunningPlayApp runningApp = new AdvancedRunningPlayApp(testDirectory)
-    PlayApp playApp = new AdvancedPlayApp()
+        then:
+        appIsRunningAndDeployed()
 
-    def cleanup() {
-        stopGradle()
-        if (runningApp.isInitialized()) {
-            appIsStopped()
+        when:
+        addNewRoute('hello')
+
+        then:
+        succeeds()
+        runningApp.playUrl('hello').text == 'hello world'
+    }
+
+    def "should reload with exception when modify scala controller"() {
+        when:
+        succeeds("runPlayBinary")
+        then:
+        appIsRunningAndDeployed()
+
+        when:
+        addBadCode()
+        then:
+        fails()
+        !executedTasks.contains('runPlayBinary')
+        errorPageHasTaskFailure("compilePlayBinaryScala")
+
+        when:
+        fixBadCode()
+        then:
+        succeeds()
+        appIsRunningAndDeployed()
+
+    }
+
+    def "should reload modified coffeescript"() {
+        when:
+        succeeds("runPlayBinary")
+
+        then:
+        appIsRunningAndDeployed()
+        !runningApp.playUrl('assets/javascripts/test.js').text.contains('Hello coffeescript')
+        !runningApp.playUrl('assets/javascripts/test.min.js').text.contains('Hello coffeescript')
+
+        when:
+        file("app/assets/javascripts/test.coffee") << '''
+message = "Hello coffeescript"
+alert message
+'''
+
+        then:
+        succeeds()
+        runningApp.playUrl('assets/javascripts/test.js').text.contains('Hello coffeescript')
+        runningApp.playUrl('assets/javascripts/test.min.js').text.contains('Hello coffeescript')
+    }
+
+    def "should detect new javascript files"() {
+        when:
+        succeeds("runPlayBinary")
+
+        then:
+        appIsRunningAndDeployed()
+
+        when:
+        file("app/assets/javascripts/helloworld.js") << '''
+var message = "Hello JS";
+'''
+
+        then:
+        succeeds()
+        runningApp.playUrl('assets/javascripts/helloworld.js').text.contains('Hello JS')
+        runningApp.playUrl('assets/javascripts/helloworld.min.js').text.contains('Hello JS')
+    }
+
+    def "should reload modified java model"() {
+        when:
+        succeeds("runPlayBinary")
+
+        then:
+        appIsRunningAndDeployed()
+        assert runningApp.playUrl().text.contains("<li>foo:1</li>")
+
+        when:
+        file("app/models/DataType.java").with {
+            text = text.replaceFirst(~/"%s:%s"/, '"Hello %s:%s !"')
         }
+
+        then:
+        succeeds()
+        assert runningApp.playUrl().text.contains("<li>Hello foo:1 !</li>")
     }
 
-    protected void addNewRoute(String route) {
-        file("conf/routes") << "\nGET     /${route}                   controllers.Application.${route}"
-        file("app/controllers/Application.scala").with {
-            text = text.replaceFirst(/(?s)\}\s*$/, """
-  def ${route} = Action {
-    Ok("${route} world")
-  }
-}
-""")
+    def "should reload twirl template"() {
+        when:
+        succeeds("runPlayBinary")
+
+        then:
+        appIsRunningAndDeployed()
+
+        when:
+        file("app/views/index.scala.html").with {
+            text = text.replaceFirst(~/Welcome to Play/, 'Welcome to Play with Gradle')
         }
+
+        then:
+        succeeds()
+        assert runningApp.playUrl().text.contains("Welcome to Play with Gradle")
     }
 
-    protected errorPageHasTaskFailure(task) {
-        def error = runningApp.playUrlError()
-        assert error.httpCode == 500
-        assert error.text.contains("Gradle Build Failure")
-        assert error.text.contains("Execution failed for task &#x27;:$task&#x27;.")
-        error
-    }
-
-    protected void addBadCode() {
-        file("conf/routes") << "\nGET     /hello                   controllers.Application.hello"
-        file("app/controllers/Application.scala").with {
-            text = text.replaceFirst(/(?s)\}\s*$/, '''
-  def hello = Action {
-    Ok("hello world")
-  }
-''') // missing closing brace
-        }
-    }
-
-    protected void fixBadCode() {
-        file("app/controllers/Application.scala") << "}"
-    }
-
-    protected void addPendingChangesHook() {
+    def "should reload with exception when task that depends on runPlayBinary fails"() {
+        given:
         buildFile << """
-            ext.pendingChanges = new java.util.concurrent.atomic.AtomicBoolean(false)
+task otherTask {
+   dependsOn 'runPlayBinary'
+   doLast {
+      // second time through this route exists
+      if (file("conf/routes").text.contains("/hello")) {
+         throw new GradleException("always fails")
+      }
+   }
+}
+"""
+        when:
+        succeeds("otherTask")
+        then:
+        appIsRunningAndDeployed()
 
-            void addPendingChangeListener(Closure action={}) {
-                def pendingChangesManager = gradle.services.get(${PendingChangesManager.canonicalName})
-                pendingChangesManager.addListener new ${SingleFirePendingChangesListener.canonicalName}({
-                    synchronized(pendingChanges) {
-                        println "Pending changes detected"
-                        pendingChanges.set(true)
-                        pendingChanges.notifyAll()
-                        action()
-                    }
-                } as ${PendingChangesListener.canonicalName})
-            }
+        when:
+        addNewRoute('hello')
 
-            void waitForPendingChanges() {
-                synchronized(pendingChanges) {
-                    while(!pendingChanges.get()) {
-                        pendingChanges.wait()
-                    }
-                }
-            }
-        """
+        then:
+        fails()
+        !executedTasks.contains('runPlayBinary')
+        errorPageHasTaskFailure("otherTask")
     }
 }
